browser-use-service/pyproject.toml
[project]
name = "browser-use-service"
version = "0.1.0"
description = "BrowserUse agent microservice (POC)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "browser-use>=0.1.40",
  "langchain-openai>=0.2.0",
  "langchain-ollama>=0.2.0",
  "langchain-core>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


